requirement.txt
certifi==2021.10.8
charset-normalizer==2.0.9
et-xmlfile==1.1.0
furl==2.1.3
idna==3.3
mysql-connector-python==8.0.27
openpyxl==3.0.9
orderedmultidict==1.0.1
protobuf==3.19.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
anyio==3.5.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
et-xmlfile==1.1.0
fastapi==0.72.0
furl==2.1.3
h11==0.13.0
idna==3.3
mysql-connector-python==8.0.27
openpyxl==3.0.9
orderedmultidict==1.0.1
protobuf==3.19.1
pydantic==1.9.0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.17.0
xmltodict==0.12.0
